plugin addition request
I made a plugin that allows you to download youtube videos. you just type -yt {URL} and it asks you if you want to download mp3 or mp4. then, after you choose, it will download the file and open you  Downloads folder. it still requires some changes, but i think it would be a great addition. if you decide to still not let it pass, i would like to know what i can improve to make it better.

i.e this is my first large flow plugin, and if i have some problems with the packaging or github workflow or something like that, i would not like it to be the reason that will stop my plugin. if i do have some of these problems, please let me know.